Move security support from main app to drivers
- No board change; continue selecting `SECURITY` config
- Security state enum moved to driver header
- Update related doc
