+++ src/app/main/Makefile.mk
@@ -18,7 +18,6 @@ app-y += pnp.c
 app-y += ps2.c
 app-y += pwm.c
 app-y += scratch.c
-app-$(CONFIG_SECURITY) += security.c
 app-y += smbus.c
 app-y += smfi.c
 app-y += stdio.c
@@ -39,10 +38,6 @@ CFLAGS+=-DLEVEL=4
 CFLAGS += -DI2C_SMBUS=$(CONFIG_I2C_SMBUS)
 # Uncomment to enable I2C debug on 0x76
 #CFLAGS+=-DI2C_DEBUGGER=0x76
-
-ifeq ($(CONFIG_SECURITY),y)
-CFLAGS+=-DCONFIG_SECURITY=1
-endif
 
 ifeq ($(CONFIG_PLATFORM_INTEL),y)
 app-y += peci.c
