+++ src/app/main/smfi.c
@@ -22,7 +22,7 @@
 #include <app/scratch.h>
 
 #if CONFIG_SECURITY
-#include <app/security.h>
+#include <drivers/security/security.h>
 #endif // CONFIG_SECURITY
 #endif // !defined(__SCRATCH__)
